Create tests prototype, add undo within a turn, fix bugs in positioning and gag functions, other small adjustments
Fix "right" position definition not updating properly before the first turn
Fix prestige sound calculations for odd-leveled cogs
Fix single drop to not include combo bonus
Make separate generate function for Battle, as well as other helper functions for accessibility
Add configuration for no miss (accuracy yet to be implemented)
Rename methods for consistent camel case naming

+++ lib/gag_collection.h
@@ -24,7 +24,7 @@ struct Gag {
     GagKind kind;
     std::string name;
     unsigned short damage;
-    float accuracy;
+    int accuracy;
     int target;
     bool prestiged;
